Installs gulp-uglify and configures JavaScript to suit it
