+++ js/jekyll-search.js
@@ -1,1 +1,1 @@
-!function e(t,n,r){function s(o,u){if(!n[o]){if(!t[o]){var a="function"==typeof require&&require;if(!u&&a)return a(o,!0);if(i)return i(o,!0);throw new Error("Cannot find module '"+o+"'")}var f=n[o]={exports:{}};t[o][0].call(f.exports,function(e){var n=t[o][1][e];return s(n?n:e)},f,f.exports,e,t,n,r)}return n[o].exports}for(var i="function"==typeof require&&require,o=0;o<r.length;o++)s(r[o]);return s}({1:[function(require,module){function JSONLoader(){function receivedResponse(xhr){return 200==xhr.status&&4==xhr.readyState}function handleResponse(xhr,callback){xhr.onreadystatechange=function(){if(receivedResponse(xhr))try{callback(null,JSON.parse(xhr.responseText))}catch(err){callback(err,null)}}}var self=this;self.load=function(location,callback){var xhr=window.XMLHttpRequest?new XMLHttpRequest:new ActiveXObject("Microsoft.XMLHTTP");xhr.open("GET",location,!0),handleResponse(xhr,callback),xhr.send()}}module.exports=new JSONLoader},{}],2:[function(require,module){function FuzzySearchStrategy(){var self=this;self.matches=function(string,crit){var regexp=new RegExp(crit.split("").join(".*?"),"gi");return!!string.match(regexp)}}module.exports=new FuzzySearchStrategy},{}],3:[function(require,module){function LiteralSearchStrategy(){var self=this;self.matches=function(string,crit){return string.toLowerCase().indexOf(crit.toLowerCase())>=0}}module.exports=new LiteralSearchStrategy},{}],4:[function(require,module){function Searcher(){function trim(string){return string.replace(/^ */,"").replace(/ *$/,"")}function getSearchStrategy(){return fuzzy?fuzzySearchStrategy:literalSearchStrategy}function findMatchesWithStrategy(data,crit,strategy){for(var matches=[],i=0;i<data.length&&matches.length<limit;i++){var obj=data[i];for(var key in obj)if(obj.hasOwnProperty(key)&&"string"==typeof obj[key]&&strategy.matches(obj[key],crit)){matches.push(obj);break}}return matches}var self=this,fuzzy=!1,limit=10,fuzzySearchStrategy=require("./SearchStrategies/fuzzy"),literalSearchStrategy=require("./SearchStrategies/literal");self.setFuzzy=function(_fuzzy){fuzzy=!!_fuzzy},self.setLimit=function(_limit){limit=parseInt(_limit,10)||limit},self.search=function(data,crit){return crit?findMatchesWithStrategy(data,trim(crit),getSearchStrategy()):[]}}module.exports=new Searcher},{"./SearchStrategies/fuzzy":2,"./SearchStrategies/literal":3}],5:[function(require,module){function Store(){function isObject(obj){return"[object Object]"==Object.prototype.toString.call(obj)}function isArray(obj){return"[object Array]"==Object.prototype.toString.call(obj)}function addObject(data){return store.push(data),data}function addArray(data){for(var added=[],i=0;i<data.length;i++)isObject(data[i])&&added.push(addObject(data[i]));return added}var self=this,store=[];self.clear=function(){store.length=0},self.get=function(){return store},self.put=function(data){return isObject(data)?addObject(data):isArray(data)?addArray(data):void 0}}module.exports=new Store},{}],6:[function(require,module){function Templater(){var self=this,templatePattern=/\{(.*?)\}/g;self.setTemplatePattern=function(newTemplatePattern){templatePattern=newTemplatePattern},self.render=function(t,data){return t.replace(templatePattern,function(match,prop){return data[prop]||match})}}module.exports=new Templater},{}],7:[function(require){!function(window){"use strict";function SimpleJekyllSearch(){function throwError(message){throw new Error("SimpleJekyllSearch --- "+message)}function validateOptions(_opt){for(var i=0;i<requiredOptions.length;i++){var req=requiredOptions[i];_opt[req]||throwError("You must specify a "+req)}}function assignOptions(_opt){for(var option in opt)opt[option]=_opt[option]||opt[option]}function isJSON(json){try{return json instanceof Object&&JSON.parse(JSON.stringify(json))}catch(e){return!1}}function registerInput(){opt.searchInput.addEventListener("keyup",function(e){render(searcher.search(store.get(),e.target.value))})}function render(results){opt.resultsContainer.innerHTML="";for(var i=0;i<results.length;i++){var result=results[i],rendered=templater.render(opt.searchResultTemplate,result);opt.resultsContainer.innerHTML+=rendered}}var self=this,requiredOptions=["searchInput","resultsContainer","dataSource"],opt={searchInput:null,resultsContainer:null,dataSource:[],searchResultTemplate:'<li><a href="{url}" title="{desc}">{title}</a></li>',noResultsText:"No results found",limit:10,fuzzy:!1};self.init=function(_opt){validateOptions(_opt),assignOptions(_opt),isJSON(opt.dataSource)?(store.put(opt.dataSource),registerInput()):JSONLoader.load(opt.dataSource,function(err,json){err?throwError("failed to get JSON ("+opt.dataSource+")"):(store.put(json),registerInput())})}}var templater=require("./Templater"),store=require("./Store"),searcher=require("./Searcher"),JSONLoader=require("./JSONLoader");window.SimpleJekyllSearch=new SimpleJekyllSearch}(window,document)},{"./JSONLoader":1,"./Searcher":4,"./Store":5,"./Templater":6}]},{},[7]);
+!function t(e,n,r){function o(i,a){if(!n[i]){if(!e[i]){var c="function"==typeof require&&require;if(!a&&c)return c(i,!0);if(u)return u(i,!0);throw new Error("Cannot find module '"+i+"'")}var f=n[i]={exports:{}};e[i][0].call(f.exports,function(t){var n=e[i][1][t];return o(n?n:t)},f,f.exports,t,e,n,r)}return n[i].exports}for(var u="function"==typeof require&&require,i=0;i<r.length;i++)o(r[i]);return o}({1:[function(t,e){function n(){function t(t){return 200==t.status&&4==t.readyState}function e(e,n){e.onreadystatechange=function(){if(t(e))try{n(null,JSON.parse(e.responseText))}catch(r){n(r,null)}}}var n=this;n.load=function(t,n){var r=window.XMLHttpRequest?new XMLHttpRequest:new ActiveXObject("Microsoft.XMLHTTP");r.open("GET",t,!0),e(r,n),r.send()}}e.exports=new n},{}],2:[function(t,e){function n(){var t=this;t.matches=function(t,e){var n=new RegExp(e.split("").join(".*?"),"gi");return!!t.match(n)}}e.exports=new n},{}],3:[function(t,e){function n(){var t=this;t.matches=function(t,e){return t.toLowerCase().indexOf(e.toLowerCase())>=0}}e.exports=new n},{}],4:[function(t,e){function n(){function e(t){return t.replace(/^ */,"").replace(/ *$/,"")}function n(){return u?a:c}function r(t,e,n){for(var r=[],o=0;o<t.length&&r.length<i;o++){var u=t[o];for(var a in u)if(u.hasOwnProperty(a)&&"string"==typeof u[a]&&n.matches(u[a],e)){r.push(u);break}}return r}var o=this,u=!1,i=10,a=t("./SearchStrategies/fuzzy"),c=t("./SearchStrategies/literal");o.setFuzzy=function(t){u=!!t},o.setLimit=function(t){i=parseInt(t,10)||i},o.search=function(t,o){return o?r(t,e(o),n()):[]}}e.exports=new n},{"./SearchStrategies/fuzzy":2,"./SearchStrategies/literal":3}],5:[function(t,e){function n(){function t(t){return"[object Object]"==Object.prototype.toString.call(t)}function e(t){return"[object Array]"==Object.prototype.toString.call(t)}function n(t){return u.push(t),t}function r(e){for(var r=[],o=0;o<e.length;o++)t(e[o])&&r.push(n(e[o]));return r}var o=this,u=[];o.clear=function(){u.length=0},o.get=function(){return u},o.put=function(o){return t(o)?n(o):e(o)?r(o):void 0}}e.exports=new n},{}],6:[function(t,e){function n(){var t=this,e=/\{(.*?)\}/g;t.setTemplatePattern=function(t){e=t},t.render=function(t,n){return t.replace(e,function(t,e){return n[e]||t})}}e.exports=new n},{}],7:[function(t){!function(e){"use strict";function n(){function t(t){throw new Error("SimpleJekyllSearch --- "+t)}function e(e){for(var n=0;n<l.length;n++){var r=l[n];e[r]||t("You must specify a "+r)}}function n(t){for(var e in p)p[e]=t[e]||p[e]}function a(t){try{return t instanceof Object&&JSON.parse(JSON.stringify(t))}catch(e){return!1}}function c(){p.searchInput.addEventListener("keyup",function(t){f(u.search(o.get(),t.target.value))})}function f(t){p.resultsContainer.innerHTML="";for(var e=0;e<t.length;e++){var n=t[e],o=r.render(p.searchResultTemplate,n);p.resultsContainer.innerHTML+=o}}var s=this,l=["searchInput","resultsContainer","dataSource"],p={searchInput:null,resultsContainer:null,dataSource:[],searchResultTemplate:'<li><a href="{url}" title="{desc}">{title}</a></li>',noResultsText:"No results found",limit:10,fuzzy:!1};s.init=function(r){e(r),n(r),a(p.dataSource)?(o.put(p.dataSource),c()):i.load(p.dataSource,function(e,n){e?t("failed to get JSON ("+p.dataSource+")"):(o.put(n),c())})}}var r=t("./Templater"),o=t("./Store"),u=t("./Searcher"),i=t("./JSONLoader");e.SimpleJekyllSearch=new n}(window,document)},{"./JSONLoader":1,"./Searcher":4,"./Store":5,"./Templater":6}]},{},[7]);
