+++ package.json
@@ -8,6 +8,7 @@
     "gulp": "^3.8.9",
     "gulp-filter": "^1.0.2",
     "gulp-rename": "^1.2.0",
+    "gulp-uglify": "^1.0.1",
     "main-bower-files": "^2.1.0"
   },
   "devDependencies": {},
